refactor(workflow): reorganize BaseNode import paths
- Moved `BaseNode` to a new `base` package for better organization.
- Updated all import statements to reflect new path structure.
- Added `__init__.py` to `base` package for proper module export.
