+++ api/core/workflow/nodes/base/__init__.py
@@ -1,0 +1,3 @@
+from .node import BaseNode
+
+__all__ = ["BaseNode"]
