+++ api/core/workflow/nodes/loop/loop_node.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from core.workflow.nodes.base_node import BaseNode
+from core.workflow.nodes.base import BaseNode
 from core.workflow.nodes.loop.entities import LoopNodeData, LoopState
 from core.workflow.utils.condition.entities import Condition
 from enums import NodeType
