+++ api/core/workflow/nodes/variable_assigner/node.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import Session
 from core.variables import SegmentType, Variable
 from core.workflow.entities.base_node_data_entities import BaseNodeData
 from core.workflow.entities.node_entities import NodeRunResult
-from core.workflow.nodes.base_node import BaseNode
+from core.workflow.nodes.base import BaseNode
 from enums import NodeType
 from extensions.ext_database import db
 from factories import variable_factory
